feat: 调整makeFailedResources
feat: 调整makeFailedResources
